+++ CHANGES.md
@@ -9,6 +9,7 @@
    when the drop-down box would contain any **compatible resource**
    [#310](https://github.com/CCI-Tools/cate-core/issues/310).
  * Only show lat/lon values when mouse is over globe [#312](https://github.com/CCI-Tools/cate-core/issues/312).
+ * Prevent a 2nd (and 3rd) invocation of the variable statistics computation
 
 ## 0.9.0-dev.4